Restore compilation with custom presets
This also assigns precise types to the constants in the minimal
and mainnet presets in order to reduce the chance of compilation
errors when custom presets are used (previously, only the custom
presets have precisely assigned types for the constants).

+++ beacon_chain/ssz/bytes_reader.nim
@@ -70,7 +70,7 @@ proc `[]`[T, U, V](s: openArray[T], x: HSlice[U, V]) {.error:
 
 template checkForForbiddenBits(ResulType: type,
                                input: openarray[byte],
-                               expectedBits: static int) =
+                               expectedBits: static int64) =
   ## This checks if the input contains any bits set above the maximum
   ## sized allowed. We only need to check the last byte to verify this:
   const bitsInLastByte = (expectedBits mod 8)
